test: 자동차 수 관련 메서드 위치 변경
